+++ index.js
@@ -1,6 +1,6 @@
 ChatLib.chat("\n&5[&b&lRFU&5] &6&lLoading...\n&cIf you dont see a loaded msg /ct reload ._.\n")
 
-import "./features/party.js";
+// import "./features/party.js";
 import "./features/generalFishing.js";
 import "./features/chFishing.js";
 import "./features/other.js";
@@ -9,9 +9,9 @@ import "./features/lavaFishing.js";
 import "./features/fishermantrack.js";
 import "./features/trophyFishing.js";
 import "./features/playerStats.js";
-import "./utils/mixenddata.js";
 import "./utils/keybinds.js";
 import "./gui/gui.js";
 import "./settings/updateSettings.js";
+import "./features/chatCommands/other.js"
 
 ChatLib.chat("\n&5[&b&lRFU&5] &2&lLoaded! \n&b&l/rfu &fto open settings\n&e&l/rfumove &fto move gui\n&9&l/rfudiscord &ffor official server\n")